Fixed port conflicts
