--- benchmarks/restify/excluded.js
@@ -1,12 +1,0 @@
-'use strict'
-
-var app = require('restify').createServer({
-  name: 'myapp',
-  version: '1.0.0'
-})
-
-app.get('/', function (req, res) {
-  res.send('content')
-})
-
-module.exports = app
