--- lib/restify.js
@@ -1,1 +1,0 @@
-export { default } from './http.js'
