+++ test/integration/express/index.js
@@ -4,9 +4,14 @@ import http from 'http'
 import express from 'express'
 import protection from '../../../index.js'
 
-// Inline minimal shim to run existing TAP-style tests under Vitest
-  if (typeof global.test !== 'function') {
+// TAP-compatible wrapper: keep Vitest `test` behaviour for normal tests
+{
+  const originalTest = global.test && global.test.bind(global)
   global.test = function (name, fn) {
+    if (!fn || fn.length === 0) {
+      if (originalTest) return originalTest(name, fn)
+      return it(name, fn)
+    }
     it(name, async () => {
       return new Promise((resolve, reject) => {
         let planned = null
@@ -47,114 +52,121 @@ function sleep (msec) {
   while (Date.now() - start < msec) {}
 }
 
-test('sends 503 when event loop is overloaded, per maxEventLoopDelay', () => {
-  return new Promise((resolve, reject) => {
-    var protect = protection('express', { maxEventLoopDelay: 1 })
+test('sends 503 when event loop is overloaded, per maxEventLoopDelay', function (t) {
+  var protect = protection('express', { maxEventLoopDelay: 1 })
 
-    var app = express()
-    app.use(protect)
-    var server = http.createServer(function (req, res) {
-      sleep(500)
-      app(req, res)
-    import protection from '../../../index.js'
+  var app = express()
+  app.use(protect)
+  app.use(function (req, res) {
+    sleep(500) // Trigger event loop delay during request
+    res.end('content')
+  })
+  var server = http.createServer(app)
 
-    server.listen(0, function () {
-      const port = server.address().port
-    if (typeof global.test !== 'function') {
-      global.test = function (name, fn) {
-        it(name, async () => {
-          return new Promise((resolve, reject) => {
-            const t = {
-              is: (a, b) => { try { expect(a).toBe(b) } catch (e) { reject(e) } },
-              same: (a, b) => { try { expect(a).toEqual(b) } catch (e) { reject(e) } },
-              ok: (v) => { try { expect(v).toBeTruthy() } catch (e) { reject(e) } },
-              fail: (msg) => reject(new Error(msg || 'fail')),
-              throws: (fnc) => { try { fnc(); reject(new Error('did not throw')) } catch (e) {} },
-              plan: function () {},
-              pass: function () {},
-              end: function () { resolve() }
-            }
-            try {
-              const maybe = fn(t)
-              if (maybe && typeof maybe.then === 'function') maybe.then(resolve, reject)
-              setTimeout(() => reject(new Error('Test did not call t.end() within timeout')), 300)
-            } catch (err) { reject(err) }
-          })
-        })
-      }
+  server.listen(0, function () {
+    const port = server.address().port
+    var req = http.get('http://localhost:' + port)
+    req.on('response', function (res) {
+      t.is(res.statusCode, 503)
+      protect.stop()
+      server.close()
+      t.end()
+    }).end()
+  })
+})
+
+test('sends 503 when heap used threshold is passed, as per maxHeapUsedBytes', function (t) {
+  var memoryUsage = process.memoryUsage
+  process.memoryUsage = function () { return { rss: 99999, heapTotal: 9999, heapUsed: 999, external: 99 } }
+  var protect = protection('express', { sampleInterval: 5, maxEventLoopDelay: 0, maxHeapUsedBytes: 40 })
+
+  var app = express()
+  app.use(protect)
+  var server = http.createServer(app)
+
+  server.listen(0, function () {
+    const port = server.address().port
+    setTimeout(function () {
+      var req = http.get('http://localhost:' + port)
+      req.on('response', function (res) {
+        t.is(res.statusCode, 503)
+        server.close()
+        protect.stop()
+        process.memoryUsage = memoryUsage
+        t.end()
+      }).end()
+    }, 6)
+  })
+})
+
+test('sends 503 when heap used threshold is passed, as per maxRssBytes', function (t) {
+  var memoryUsage = process.memoryUsage
+  process.memoryUsage = function () {
+    return {
+      rss: 99999,
+      heapTotal: 9999,
+      heapUsed: 999,
+      external: 99
     }
-    server.listen(0, function () {
-      const port = server.address().port
+  }
+  var protect = protection('express', {
+    sampleInterval: 5,
+    maxEventLoopDelay: 0,
+    maxRssBytes: 40
+  })
 
-        var req = http.get('http://localhost:' + port)
-        req.on('response', function (res) {
-          try {
-            expect(res.statusCode).toBe(503)
-            server.close()
-            protect.stop()
-            process.memoryUsage = memoryUsage
-            resolve()
-          } catch (err) { reject(err) }
-        }).on('error', reject).end()
-      }, 6)
-    })
+  var app = express()
+  app.use(protect)
+  var server = http.createServer(app)
+
+  server.listen(0, function () {
+    setTimeout(function () {
+      const port = server.address().port
+      var req = http.get('http://localhost:' + port)
+      req.on('response', function (res) {
+        t.is(res.statusCode, 503)
+        server.close()
+        protect.stop()
+        process.memoryUsage = memoryUsage
+        t.end()
+      }).end()
+    }, 6)
   })
 })
 
-test('sends 503 when heap used threshold is passed, as per maxRssBytes', () => {
-  return new Promise((resolve, reject) => {
-    var memoryUsage = process.memoryUsage
-    process.memoryUsage = function () { return { rss: 99999, heapTotal: 9999, heapUsed: 999, external: 99 } }
-    var protect = protection('express', { sampleInterval: 5, maxEventLoopDelay: 0, maxRssBytes: 40 })
-
-    var app = express()
-    app.use(protect)
-    var server = http.createServer(app)
-
-    server.listen(0, function () {
-      const port = server.address().port
-      setTimeout(function () {
-        var req = http.get('http://localhost:' + port)
-        req.on('response', function (res) {
-          try {
-            expect(res.statusCode).toBe(503)
-            server.close()
-            protect.stop()
-            process.memoryUsage = memoryUsage
-            resolve()
-          } catch (err) { reject(err) }
-        }).on('error', reject).end()
-      }, 6)
-    })
+test('sends Retry-After header as per clientRetrySecs', function (t) {
+  var memoryUsage = process.memoryUsage
+  process.memoryUsage = function () {
+    return {
+      rss: 99999,
+      heapTotal: 9999,
+      heapUsed: 999,
+      external: 99
+    }
+  }
+  var protect = protection('express', {
+    sampleInterval: 5,
+    maxEventLoopDelay: 0,
+    maxRssBytes: 40,
+    clientRetrySecs: 22
   })
-})
 
-test('sends Retry-After header as per clientRetrySecs', () => {
-  return new Promise((resolve, reject) => {
-    var memoryUsage = process.memoryUsage
-    process.memoryUsage = function () { return { rss: 99999, heapTotal: 9999, heapUsed: 999, external: 99 } }
-    var protect = protection('express', { sampleInterval: 5, maxEventLoopDelay: 0, maxRssBytes: 40, clientRetrySecs: 22 })
+  var app = express()
+  app.use(protect)
+  var server = http.createServer(app)
 
-    var app = express()
-    app.use(protect)
-    var server = http.createServer(app)
-
-    server.listen(0, function () {
-      const port = server.address().port
-      setTimeout(function () {
-        var req = http.get('http://localhost:' + port)
-        req.on('response', function (res) {
-          try {
-            expect(res.statusCode).toBe(503)
-            expect(res.headers['retry-after']).toBe('22')
-            server.close()
-            protect.stop()
-            process.memoryUsage = memoryUsage
-            resolve()
-          } catch (err) { reject(err) }
-        }).on('error', reject).end()
-      }, 6)
-    })
+  server.listen(0, function () { const port = server.address().port
+    setTimeout(function () {
+      var req = http.get('http://localhost:' + port)
+      req.on('response', function (res) {
+        t.is(res.statusCode, 503)
+        t.is(res.headers['retry-after'], '22')
+        server.close()
+        protect.stop()
+        process.memoryUsage = memoryUsage
+        t.end()
+      }).end()
+    }, 6)
   })
 })
 
@@ -168,7 +180,7 @@ test('does not set Retry-After header when clientRetrySecs is 0', function (t) {
       external: 99
     }
   }
-
+  var protect = protection('express', {
     sampleInterval: 5,
     maxEventLoopDelay: 0,
     maxRssBytes: 40,
@@ -179,27 +191,22 @@ test('does not set Retry-After header when clientRetrySecs is 0', function (t) {
   app.use(protect)
   var server = http.createServer(app)
 
-  return new Promise((resolve, reject) => {
-    server.listen(0, function () {
-      const port = server.address().port
-      setTimeout(function () {
-        var req = http.get('http://localhost:' + port)
-        req.on('response', function (res) {
-          try {
-            expect(res.statusCode).toBe(503)
-            expect('retry-after' in res.headers).toBe(false)
-            server.close()
-            protect.stop()
-            process.memoryUsage = memoryUsage
-            resolve()
-          } catch (err) { reject(err) }
-        }).on('error', reject).end()
-      }, 6)
-    })
+  server.listen(0, function () { const port = server.address().port
+    setTimeout(function () {
+      var req = http.get('http://localhost:' + port)
+      req.on('response', function (res) {
+        t.is(res.statusCode, 503)
+        t.is('retry-after' in res.headers, false)
+        server.close()
+        protect.stop()
+        process.memoryUsage = memoryUsage
+        t.end()
+      }).end()
+    }, 6)
   })
 })
 
-test('errorPropagationMode:false (default)', () => {
+test('errorPropagationMode:false (default)', function (t) {
   var memoryUsage = process.memoryUsage
   process.memoryUsage = function () {
     return {
@@ -218,31 +225,27 @@ test('errorPropagationMode:false (default)', () => {
 
   var app = express()
   app.use(protect)
-  app.use(function () {
-    throw new Error('should not be called')
+  app.use(function (req, res) {
+    t.fail() // should never be called
+    res.end('content')
   })
   var server = http.createServer(app)
 
-  return new Promise((resolve, reject) => {
-    server.listen(0, function () {
-      const port = server.address().port
-      setTimeout(function () {
-        var req = http.get('http://localhost:' + port)
-        req.on('response', function (res) {
-          try {
-            expect(res.statusCode).toBe(503)
-            server.close()
-            protect.stop()
-            process.memoryUsage = memoryUsage
-            resolve()
-          } catch (err) { reject(err) }
-        }).on('error', reject).end()
-      }, 6)
-    })
+  server.listen(0, function () { const port = server.address().port
+    setTimeout(function () {
+      var req = http.get('http://localhost:' + port)
+      req.on('response', function (res) {
+        t.is(res.statusCode, 503)
+        server.close()
+        protect.stop()
+        process.memoryUsage = memoryUsage
+        t.end()
+      }).end()
+    }, 6)
   })
 })
 
-test('errorPropagationMode:true', () => {
+test('errorPropagationMode:true', function (t) {
   var memoryUsage = process.memoryUsage
   process.memoryUsage = function () {
     return {
@@ -262,36 +265,27 @@ test('errorPropagationMode:true', () => {
   var app = express()
   app.use(protect)
   app.use(function (err, req, res, next) {
-    try {
-      expect(err).toBeTruthy()
-      expect(err.statusCode).toBe(503)
-    } catch (e) {
-      // still respond
-    }
+    t.ok(err)
+    t.is(err.statusCode, 503)
     res.end('err message')
   })
   var server = http.createServer(app)
 
-  return new Promise((resolve, reject) => {
-    server.listen(0, function () {
-      const port = server.address().port
-      setTimeout(function () {
-        var req = http.get('http://localhost:' + port)
-        req.on('response', function (res) {
-          try {
-            expect(res.statusCode).toBe(503)
-            server.close()
-            protect.stop()
-            process.memoryUsage = memoryUsage
-            resolve()
-          } catch (err) { reject(err) }
-        }).on('error', reject).end()
-      }, 6)
-    })
+  server.listen(0, function () { const port = server.address().port
+    setTimeout(function () {
+      var req = http.get('http://localhost:' + port)
+      req.on('response', function (res) {
+        t.is(res.statusCode, 503)
+        server.close()
+        protect.stop()
+        process.memoryUsage = memoryUsage
+        t.end()
+      }).end()
+    }, 6)
   })
 })
 
-test('in default mode, production:false leads to high detail client response message', () => {
+test('in default mode, production:false leads to high detail client response message', function (t) {
   var memoryUsage = process.memoryUsage
   process.memoryUsage = function () {
     return {
@@ -313,30 +307,25 @@ test('in default mode, production:false leads to high detail client response mes
   app.use(protect)
   var server = http.createServer(app)
 
-  return new Promise((resolve, reject) => {
-    server.listen(0, function () {
-      const port = server.address().port
-      setTimeout(function () {
-        var req = http.get('http://localhost:' + port)
-        req.on('response', function (res) {
-          try {
-            expect(res.statusCode).toBe(503)
-            res.once('data', function (msg) {
-              msg = msg.toString()
-              expect(msg).toBe('Server experiencing heavy load: (rss)')
-              server.close()
-              protect.stop()
-              process.memoryUsage = memoryUsage
-              resolve()
-            })
-          } catch (err) { reject(err) }
-        }).on('error', reject).end()
-      }, 6)
-    })
+  server.listen(0, function () { const port = server.address().port
+    setTimeout(function () {
+      var req = http.get('http://localhost:' + port)
+      req.on('response', function (res) {
+        t.is(res.statusCode, 503)
+        res.once('data', function (msg) {
+          msg = msg.toString()
+          t.is(msg, 'Server experiencing heavy load: (rss)')
+          server.close()
+          protect.stop()
+          process.memoryUsage = memoryUsage
+          t.end()
+        })
+      }).end()
+    }, 6)
   })
 })
 
-test('in default mode, production:true leads to standard 503 client response message', () => {
+test('in default mode, production:true leads to standard 503 client response message', function (t) {
   var memoryUsage = process.memoryUsage
   process.memoryUsage = function () {
     return {
@@ -358,30 +347,25 @@ test('in default mode, production:true leads to standard 503 client response mes
   app.use(protect)
   var server = http.createServer(app)
 
-  return new Promise((resolve, reject) => {
-    server.listen(0, function () {
-      const port = server.address().port
-      setTimeout(function () {
-        var req = http.get('http://localhost:' + port)
-        req.on('response', function (res) {
-          try {
-            expect(res.statusCode).toBe(503)
-            res.once('data', function (msg) {
-              msg = msg.toString()
-              expect(msg).toBe('Service Unavailable')
-              server.close()
-              protect.stop()
-              process.memoryUsage = memoryUsage
-              resolve()
-            })
-          } catch (err) { reject(err) }
-        }).on('error', reject).end()
-      }, 6)
-    })
+  server.listen(0, function () { const port = server.address().port
+    setTimeout(function () {
+      var req = http.get('http://localhost:' + port)
+      req.on('response', function (res) {
+        t.is(res.statusCode, 503)
+        res.once('data', function (msg) {
+          msg = msg.toString()
+          t.is(msg, 'Service Unavailable')
+          server.close()
+          protect.stop()
+          process.memoryUsage = memoryUsage
+          t.end()
+        })
+      }).end()
+    }, 6)
   })
 })
 
-test('in errorPropagationMode production:false sets expose:true on error object', () => {
+test('in errorPropagationMode production:false sets expose:true on error object', function (t) {
   var memoryUsage = process.memoryUsage
   process.memoryUsage = function () {
     return {
@@ -402,34 +386,27 @@ test('in errorPropagationMode production:false sets expose:true on error object'
   var app = express()
   app.use(protect)
   app.use(function (err, req, res, next) {
-    try {
-      expect(err).toBeTruthy()
-      expect(err.expose).toBe(true)
-    } catch (e) {}
+    t.ok(err)
+    t.is(err.expose, true)
     res.end('err message')
   })
   var server = http.createServer(app)
 
-  return new Promise((resolve, reject) => {
-    server.listen(0, function () {
-      const port = server.address().port
-      setTimeout(function () {
-        var req = http.get('http://localhost:' + port)
-        req.on('response', function (res) {
-          try {
-            expect(res.statusCode).toBe(503)
-            server.close()
-            protect.stop()
-            process.memoryUsage = memoryUsage
-            resolve()
-          } catch (err) { reject(err) }
-        }).on('error', reject).end()
-      }, 6)
-    })
+  server.listen(0, function () { const port = server.address().port
+    setTimeout(function () {
+      var req = http.get('http://localhost:' + port)
+      req.on('response', function (res) {
+        t.is(res.statusCode, 503)
+        server.close()
+        protect.stop()
+        process.memoryUsage = memoryUsage
+        t.end()
+      }).end()
+    }, 6)
   })
 })
 
-test('in errorPropagationMode production:true sets expose:false on error object', () => {
+test('in errorPropagationMode production:true sets expose:false on error object', function (t) {
   var memoryUsage = process.memoryUsage
   process.memoryUsage = function () {
     return {
@@ -450,30 +427,23 @@ test('in errorPropagationMode production:true sets expose:false on error object'
   var app = express()
   app.use(protect)
   app.use(function (err, req, res, next) {
-    try {
-      expect(err).toBeTruthy()
-      expect(err.expose).toBe(false)
-    } catch (e) {}
+    t.ok(err)
+    t.is(err.expose, false)
     res.end('err message')
   })
   var server = http.createServer(app)
 
-  return new Promise((resolve, reject) => {
-    server.listen(0, function () {
-      const port = server.address().port
-      setTimeout(function () {
-        var req = http.get('http://localhost:' + port)
-        req.on('response', function (res) {
-          try {
-            expect(res.statusCode).toBe(503)
-            server.close()
-            protect.stop()
-            process.memoryUsage = memoryUsage
-            resolve()
-          } catch (err) { reject(err) }
-        }).on('error', reject).end()
-      }, 6)
-    })
+  server.listen(0, function () { const port = server.address().port
+    setTimeout(function () {
+      var req = http.get('http://localhost:' + port)
+      req.on('response', function (res) {
+        t.is(res.statusCode, 503)
+        server.close()
+        protect.stop()
+        process.memoryUsage = memoryUsage
+        t.end()
+      }).end()
+    }, 6)
   })
 })
 
@@ -495,12 +465,14 @@ test('resumes usual operation once load pressure is reduced under threshold', fu
 
   var app = express()
   app.use(protect)
-  app.get('/', function (req, res) { res.end('content') })
+  app.use(function (req, res) {
+    res.end('content')
+  })
   var server = http.createServer(app)
 
-    server.listen(0, function () { const port = server.address().port
-      setTimeout(function () {
-        var req = http.get('http://localhost:' + port)
+  server.listen(0, function () { const port = server.address().port
+    setTimeout(function () {
+      var req = http.get('http://localhost:' + port)
       req.on('response', function (res) {
         t.is(res.statusCode, 503)
         process.memoryUsage = function () {
@@ -537,8 +509,9 @@ test('if logging option is a string, when overloaded, writes log message using r
   }
   var protect = protection('express', {
     sampleInterval: 5,
-    maxEventLoopDelay: 0,
+    maxEventLoopDelay: 1,
     maxRssBytes: 40,
+    maxHeapUsedBytes: 40,
     logging: 'warn'
   })
 
@@ -546,7 +519,7 @@ test('if logging option is a string, when overloaded, writes log message using r
   app.use(function (req, res, next) {
     req.log = {
       warn: function (msg) {
-        t.is(msg, 'Server experiencing heavy load: (rss)')
+        t.is(msg, 'Server experiencing heavy load: (event loop, heap, rss)')
         server.close()
         protect.stop()
         process.memoryUsage = memoryUsage
@@ -556,11 +529,11 @@ test('if logging option is a string, when overloaded, writes log message using r
     next()
   })
   app.use(protect)
-  app.get('/', function (req, res) { res.end('content') })
   var server = http.createServer(app)
 
   server.listen(0, function () { const port = server.address().port
     setTimeout(function () {
+      sleep(500)
       http.get('http://localhost:' + port).end()
     }, 6)
   })
@@ -591,7 +564,6 @@ test('if logging option is a function, when overloaded calls the function with h
 
   var app = express()
   app.use(protect)
-  app.get('/', function (req, res) { res.end('content') })
   var server = http.createServer(app)
 
   server.listen(0, function () { const port = server.address().port
@@ -629,7 +601,9 @@ test('if logStatsOnReq is true and if logging option is a string, writes log mes
     next()
   })
   app.use(protect)
-  app.get('/', function (req, res) { res.end('content') })
+  app.use(function (req, res) {
+    res.end('content')
+  })
   var server = http.createServer(app)
 
   server.listen(0, function () { const port = server.address().port
@@ -661,11 +635,15 @@ test('if logStatsOnReq is true and logging option is a function, calls the funct
 
   var app = express()
   app.use(protect)
-  app.get('/', function (req, res) { res.end('content') })
+  app.use(function (req, res) {
+    res.end('content')
+  })
   var server = http.createServer(app)
 
-    server.listen(0, function () { const port = server.address().port
-      setTimeout(function () {
-        var req = http.get('http://localhost:' + port)
+  server.listen(0, function () {
+    setTimeout(function () {
+      const port = server.address().port
+      http.get('http://localhost:' + port).end()
+    }, 6)
   })
 })
